Add a filter stage to avro-kafka reader
Filters are pluggable via the .conf file. Filters are attached to the input kafka
topics via a .conf-specified regex pattern.

Filters are capable of:
. changing key and value types (don't have to follow the avro types directly)
. adding, combining, removing columns
. transforming data values, e.g. scaling
. rejecting out of range values

The filter stage (and any plugged in filters) is not tied to avro or
kafka specifically. Hopefully this should make filtering portable to the bulk
and csv file loaders.

Related support features
. add utility for pretty-printing avro records
. add test fixture to test stream loading avro records
. simplify type recovery in stream loading path
. improve error reporting in handling of supported cassandra column types
. bump to current version of RX (which fixes an issue connecting Subjects to Observables)
. rename to a type tagged slice of input data for a column as: TaggedSlice